add msg2 support to bp+. factor out amount recovery related classes from logic class
