+++ src/blsct/range_proof/bulletproofs/range_proof_logic.h
@@ -10,55 +10,15 @@
 
 #include <blsct/arith/elements.h>
 #include <blsct/range_proof/common.h>
+#include <blsct/range_proof/bulletproofs/amount_recovery_request.h>
+#include <blsct/range_proof/bulletproofs/amount_recovery_result.h>
 #include <blsct/range_proof/bulletproofs/range_proof_with_transcript.h>
+#include <blsct/range_proof/bulletproofs/recovered_amount.h>
 #include <consensus/amount.h>
 #include <ctokens/tokenid.h>
 #include <hash.h>
 
 namespace bulletproofs {
-
-template <typename T>
-struct AmountRecoveryRequest {
-    using Scalar = typename T::Scalar;
-    using Point = typename T::Point;
-    using Points = Elements<Point>;
-
-    size_t id;
-    Scalar x;
-    Scalar z;
-    Points Vs;
-    Points Ls;
-    Points Rs;
-    Scalar mu;
-    Scalar tau_x;
-    Point nonce;
-
-    static AmountRecoveryRequest<T> of(RangeProof<T>& proof, Point& nonce);
-};
-
-template <typename T>
-struct RecoveredAmount {
-    using Scalar = typename T::Scalar;
-
-    RecoveredAmount(
-        const size_t& id,
-        const CAmount& amount,
-        const Scalar& gamma,
-        const std::string& message) : id{id}, amount{amount}, gamma{gamma}, message{message} {}
-
-    size_t id;
-    CAmount amount;
-    Scalar gamma;
-    std::string message;
-};
-
-template <typename T>
-struct AmountRecoveryResult {
-    bool is_completed; // does not mean recovery success
-    std::vector<RecoveredAmount<T>> amounts;
-
-    static AmountRecoveryResult<T> failure();
-};
 
 // implementation of range proof algorithms described in:
 // https://eprint.iacr.org/2017/1066.pdf
